writer

+++ SOLVER/src/preloop/nrfield/wisdom/NuWisdom.cpp
@@ -36,7 +36,7 @@ void NuWisdom::writeToFile(const std::string &fname) const {
         std::vector<size_t> dims;
         dims.push_back(mRTree.size());
         dims.push_back(4);
-        ncw.defineVariable("axisem3d_wisdom", dims, (double)0.);
+        ncw.defineVariable("axisem3d_wisdom", dims, 0.);
         ncw.writeVariableWhole("axisem3d_wisdom", data);
         ncw.close();
     }
